% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickhist.R
\name{quick_hist}
\alias{quick_hist}
\title{Quick histogram}
\usage{
quick_hist(data, x, binwidth)
}
\arguments{
\item{x}{Numeric column to be plotted.}

\item{binwidth}{Number to specify bin width of the plot.}
}
\value{
A ggplot of a histogram with specified bin widths.
}
\description{
This function can be used to create quick histogram plots with specified bin widths. It is useful for comparing which bin width is best to use for a variable.
}
